message action & chat

+++ pkg/errno/errno.go
@@ -10,7 +10,7 @@ import (
 )
 
 type ErrNo struct {
-	ErrCode int64
+	ErrCode int32
 	ErrMsg  string
 }
 
@@ -18,7 +18,7 @@ func (e ErrNo) Error() string {
 	return fmt.Sprintf("err_code=%d, err_msg=%s", e.ErrCode, e.ErrMsg)
 }
 
-func NewErrNo(code int64, msg string) ErrNo {
+func NewErrNo(code int32, msg string) ErrNo {
 	return ErrNo{
 		ErrCode: code,
 		ErrMsg:  msg,
@@ -31,12 +31,12 @@ func (e ErrNo) WithMessage(msg string) ErrNo {
 }
 
 var (
-	Success                = NewErrNo(int64(user.ErrCode_SuccessCode), "Success")
-	ServiceErr             = NewErrNo(int64(user.ErrCode_ServiceErrCode), "Service is unable to start successfully")
-	ParamErr               = NewErrNo(int64(user.ErrCode_ParamErrCode), "Wrong Parameter has been given")
-	UserAlreadyExistErr    = NewErrNo(int64(user.ErrCode_UserAlreadyExistErrCode), "User already exists")
-	AuthorizationFailedErr = NewErrNo(int64(user.ErrCode_AuthorizationFailedErrCode), "Authorization failed")
-	MessageIsNullExistErr  = NewErrNo(int64(message.ErrCode_MessageIsNullErrCode ), "Message is Null")
+	Success                = NewErrNo(int32(user.ErrCode_SuccessCode), "Success")
+	ServiceErr             = NewErrNo(int32(user.ErrCode_ServiceErrCode), "Service is unable to start successfully")
+	ParamErr               = NewErrNo(int32(user.ErrCode_ParamErrCode), "Wrong Parameter has been given")
+	UserAlreadyExistErr    = NewErrNo(int32(user.ErrCode_UserAlreadyExistErrCode), "User already exists")
+	AuthorizationFailedErr = NewErrNo(int32(user.ErrCode_AuthorizationFailedErrCode), "Authorization failed")
+	MessageIsNullExistErr  = NewErrNo(int32(message.ErrCode_MessageIsNullErrCode ), "Message is Null")
 )
 
 // ConvertErr convert error to Errno
